Combine ProjectTable and ProjectList components
